+++ src/main/java/com/supermartijn642/landmines/generators/LandmineRecipeGenerator.java
@@ -23,7 +23,7 @@ public class LandmineRecipeGenerator extends RecipeGenerator {
     @Override
     public void generate(){
         // Pressure plates tag
-        TagKey<Item> pressurePlates = TagKey.create(Registries.ITEM, new ResourceLocation("landmines", "stone_pressure_plates"));
+        TagKey<Item> pressurePlates = TagKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath("landmines", "stone_pressure_plates"));
 
         // Explosion
         this.shaped(LandmineType.EXPLOSIVE.getItem())