+++ changelog.md
@@ -1,2 +1,4 @@
+## Update to Minecraft 26.1
+
 ### Landmines 1.0.7
 - Initial release of Landmines for Minecraft 1.21.11